requirements.txt
python-telegram-bot
python-telegram-bot
langchain
langchain-ollama
faiss-cpu
sentence-transformers
